--- loom.time/Constants.cs
@@ -1,8 +1,0 @@
-﻿namespace loom.time
-{
-	public static class Constants
-	{
-		public static string Username = "Xamarin";
-		public static string Password = "password";
-	}
-}
