--- loom.time/SignUpPage.xaml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" 
-			 xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
-			 x:Class="loom.time.SignUpPage"
-			 Title="Sign Up">
-    
-	<ContentPage.Content>
-		<StackLayout VerticalOptions="StartAndExpand">
-			<Label Text="Username" />
-			<Entry x:Name="usernameEntry" Placeholder="username" />
-			<Label Text="Password" />
-			<Entry x:Name="passwordEntry" IsPassword="true" />
-			<Label Text="Email address" />
-			<Entry x:Name="emailEntry" />
-			<Button Text="Sign Up" Clicked="OnSignUpButtonClicked" />
-			<Label x:Name="messageLabel" />
-		</StackLayout>
-	</ContentPage.Content>
-    
-</ContentPage>
