--- loom.time/SignUpPage.xaml.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Linq;
-using Xamarin.Forms;
-
-namespace loom.time
-{
-	public partial class SignUpPage : ContentPage
-	{
-		public SignUpPage ()
-		{
-			InitializeComponent ();
-		}
-
-		async void OnSignUpButtonClicked (object sender, EventArgs e)
-		{
-			var user = new User () {
-				Username = usernameEntry.Text,
-				Password = passwordEntry.Text,
-				Email = emailEntry.Text
-			};
-
-			// Sign up logic goes here
-
-			var signUpSucceeded = AreDetailsValid (user);
-			if (signUpSucceeded) {
-				var rootPage = Navigation.NavigationStack.FirstOrDefault ();
-				if (rootPage != null) {
-					App.IsUserLoggedIn = true;
-					Navigation.InsertPageBefore (new Views.MainTabbedPage (), Navigation.NavigationStack.First ());
-					await Navigation.PopToRootAsync ();
-				}
-			} else {
-				messageLabel.Text = "Sign up failed";
-			}
-		}
-
-		bool AreDetailsValid (User user)
-		{
-			return (!string.IsNullOrWhiteSpace (user.Username) && !string.IsNullOrWhiteSpace (user.Password) && !string.IsNullOrWhiteSpace (user.Email) && user.Email.Contains ("@"));
-		}
-	}
-}
