--- loom.time/User.cs
@@ -1,11 +1,0 @@
-﻿namespace loom.time
-{
-	public class User
-	{
-		public string Username { get; set; }
-
-		public string Password { get; set; }
-
-		public string Email { get; set; }
-	}
-}
