--- loom.time/Views/Login.xaml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" 
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
-             x:Class="loom.time.Views.Login" 
-             Icon="login.png" 
-             Title="Login">
-      
-    <ContentPage.Content>
-         
-    <FlexLayout Direction="Column"
-                AlignItems="Center"
-                JustifyContent="Start">
-            
-            <Entry x:Name="EntryStaff"
-                   Placeholder="Enter Staff Number"
-                   FontFamily="Verdana"
-                   WidthRequest="250"
-                   HeightRequest="50"
-                   Keyboard="Numeric"
-                   MaxLength="4"
-                   FontSize="Large"
-                   Margin="0,50,0,0"
-                   HorizontalTextAlignment="Center"/>
-            
-            <Button x:Name="LoginButton"
-                    Text="Login"
-                    Font="Verdana"
-                    TextColor="Black"
-                    WidthRequest="250"
-                    HeightRequest="50"
-                    FontSize="Large"
-                    Margin="0,20,0,0"
-                    BackgroundColor="White"
-                    BorderColor="#b2b2b2"
-                    BorderWidth="1"
-                    Pressed="Handle_Pressed"/>
-            
-            <Label x:Name="StaffName"
-                   Text=""
-                   Font="Verdana"
-                   HorizontalTextAlignment="Center"
-                   BackgroundColor="White"
-                   VerticalTextAlignment="Center"
-                   FontSize="Large"
-                   IsVisible="false"
-                   Margin="0,20,0,0"
-                   WidthRequest="250"
-                   HeightRequest="50"/>
-                
-        </FlexLayout>
-    </ContentPage.Content>
-    
-</ContentPage>
-
-
