+++ loom.time/Views/MainTabbedPageCS.cs
@@ -28,7 +28,7 @@ namespace loom.time.Views
 		async void OnLogoutButtonClicked (object sender, EventArgs e)
 		{
 			App.IsUserLoggedIn = false;
-			Navigation.InsertPageBefore (new LoginPageCS (), this);
+			Navigation.InsertPageBefore (new LoginPage (), this);
 			await Navigation.PopAsync ();
 		}
 	}
