+++ loom.time/Views/SignUpPage.xaml.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Linq;
+using Xamarin.Forms;
+
+namespace loom.time.Views
+{
+	public partial class SignUpPage : ContentPage
+	{
+		public SignUpPage ()
+
+		{
+            BackgroundImage = "background.jpg";
+            InitializeComponent ();
+		}
+
+		async void OnSignUpButtonClicked (object sender, EventArgs e)
+		{
+
+            //Hier den Test
+
+		}
+
+
+	}
+}
