--- loom.time/Views/Today.xaml
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" x:Class="loom.time.Views.Today">
-    <ContentPage.Content>
-    </ContentPage.Content>
-</ContentPage>
