--- loom.time/Views/Today.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-using Xamarin.Forms;
-
-namespace loom.time.Views
-{
-    public partial class Today : ContentPage
-    {
-        public Today()
-        {
-            InitializeComponent();
-        }
-    }
-}
